DjangoProject/requirements.txt
# REQUIREMENTS.TXT FOR ENHANCED HELPDESK SYSTEM
# Install all packages: pip install -r requirements.txt --break-system-packages

Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-auto-logout>=0.5.1
django-import-export>=4.1.1
python-dotenv>=1.0.1

Pillow>=10.0.0

matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0

openpyxl>=3.1.0  
reportlab>=4.0.0 
python-docx>=0.8.11  

django-anymail>=10.0  

django-ratelimit>=4.0.0  
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
drf-spectacular>=0.26.0  

python-dateutil>=2.8.2
pytz>=2023.3

gunicorn>=21.2.0  
whitenoise>=6.5.0 

django-debug-toolbar>=4.1.0
django-extensions>=3.2.3

pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.2.0
